hv: enable lapic pass-through for privileged mode
This commit enabled lapic pass-through for privileged mode. Similar to the partition mode, the vlapic is
working under the xapic mode at first, and the x2apic capability is exported to privileged vm. Only when x2apic mode
is enabled, lapic is passed through, because the physical LAPICs are under x2apic mode.

Main changes includes:
- add "is_privileged" to struct acrn_vm indicate that a VM is privileged;
- add is_lapic_pt() to check if a vcpu is created with lapic pt or not, to combine
  codes of partition mode and privileged mode, including:
  - reuse the irq delievery function and rename it to dispatch_interrupt_lapic_pt();
  - reuse switch_apicv_mode_x2apic();
  - reuse ICR handling codes to avoid malicious IPI;
- add is_privil_mode() to check if a vcpu is under privileged mode, to avoid changes for normal VM;
- for privileged vm, irq is always disabled under root mode.

Signed-off-by: Yan, Like <like.yan@intel.com>