dm: Don't notify io completion done when poweroff privileged uos
For privileged vm, lapic is passthrough to uos and we cannot kick off
the vcpu to hypervisor through IPI. Here we cannot notify the VHM/hypervisor
on the completion at this point if the UOS is in poweroff mode. Otherwise,
it will re-enter the uos and we will not have a chance to pause the vcpu.
we will mark it as done when pausing vcpu.

Signed-off-by: Kaige Fu <kaige.fu@intel.com>