Doc: Minor fix for rt_linux.rst
Actually, the link in this doc for real-time patch is for latest version of patch instead of
the 4.19.31-rt18 and it has been updated to the latest one. We can't download the 4.19.31-rt18 version
of real-time patch through it now. This patch fix it by using the permanent one.

BTW, we should execute the "cd acrn-kernel" before the "git checkout acrn-2019w17.4-160000p". This patch fixes it.

Signed-off-by: Kaige Fu <kaige.fu@intel.com>